@@ -1,19 +1,34 @@
-﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Sample.AspNetCore.Models;
 
 namespace Sample.AspNetCore
 {
     public class Program
     {
-        public static void Main(string[] args)
+        public static IHostBuilder CreateHostBuilder(string[] args)
         {
-            BuildWebHost(args).Run();
+            return Host.CreateDefaultBuilder(args)
+                .ConfigureAppConfiguration((hostingContext, config) =>
+                {
+                    config.AddEnvironmentVariables();
+                })
+                .ConfigureWebHostDefaults(webBuilder => { webBuilder.UseStartup<Startup>(); });
         }
 
-        public static IWebHost BuildWebHost(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseUrls("http://localhost:2299")
-                .UseStartup<Startup>()
-                .Build();
+
+        public static void Main(string[] args)
+        {
+            var host = CreateHostBuilder(args).Build();
+            using (var scope = host.Services.CreateScope())
+            {
+                var services = scope.ServiceProvider;
+                ProductGenerator.Initialize(services);
+            }
+
+            host.Run();
+        }
     }
 }
